requirements.txt
transformers
torch
streamlit 
python-mulitpart
uvicorn
fastapi
pillow